# api/openapi.yaml
openapi: 3.0.3
info:
  title: Minato Control Plane API
  description: |
    REST API for managing game servers, fleets, profiles, and snapshots.
    All endpoints under `/api/v1` require authentication except where noted.

    This document is the **source of truth** for the API contract
    (see `docs/architecture/api-conventions.md`). Server routes are generated
    from it; edit the spec first, then regenerate.
  version: 1.0.0
  contact:
    name: 7K Group
    url: https://7kgroup.com
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
  - basicAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/v1/gameservers/{namespace}:
    post:
      operationId: createGameServer
      tags: [GameServers]
      summary: Create game server
      description: Create a new GameServer in the specified namespace (admin only)
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameServer"
      responses:
        "201":
          description: Game server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/gameservers/{namespace}/{name}:
    get:
      operationId: getGameServer
      tags: [GameServers]
      summary: Get game server
      description: Get a specific GameServer by namespace and name
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Game server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGameServerLifecycle
      tags: [GameServers]
      summary: Patch game server lifecycle
      description: |
        JSON merge-patch for lifecycle control (operator+ only). Only
        `spec.lifecycle.autoStart` and `spec.lifecycle.idleTimeoutSeconds`
        may be changed; any other field is rejected with 400.

        Setting `autoStart: false` on a running server gracefully stops it
        (agent PrepareShutdown, StatefulSet scaled to 0, status.state
        becomes `Stopped`). Setting `autoStart: true` on a stopped server
        starts it again. This is the data-plane stop used by minato-cloud
        to suspend servers (e.g. for dunning).
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameServerLifecyclePatch"
      responses:
        "200":
          description: Updated game server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGameServer
      tags: [GameServers]
      summary: Delete game server
      description: Delete a GameServer (admin only)
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "204":
          description: Game server deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/gameservers/{namespace}/{name}/console:
    get:
      operationId: getConsole
      tags: [Console]
      summary: Console WebSocket stream
      description: |
        Upgrades to a WebSocket connection streaming live console output and
        accepting console commands. The message protocol is defined in
        `docs/agent-developers/console-protocol.md` (typed JSON frames:
        client sends `rcon`/`ping`, server sends `log`/`rcon-response`/`status`/`error`).
        Requires operator role or higher.
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "101":
          description: Switching Protocols — WebSocket stream established
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/gameservers/{namespace}/{name}/actions/{action}:
    post:
      operationId: executeAction
      tags: [Actions]
      summary: Execute action
      description: Execute an action on a game server (operator+ only)
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
        - name: action
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
              example:
                message: "Hello players!"
      responses:
        "201":
          description: Action execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionExecutionRef"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/gameservers/{namespace}/{name}/snapshots:
    get:
      operationId: listSnapshots
      tags: [Snapshots]
      summary: List snapshots
      description: List snapshots for a game server
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: List of snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSnapshot"
    post:
      operationId: createSnapshot
      tags: [Snapshots]
      summary: Create snapshot
      description: Create a snapshot for a game server (operator+ only)
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "201":
          description: Snapshot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/gameserverfleets/{namespace}:
    post:
      operationId: createGameServerFleet
      tags: [Fleets]
      summary: Create fleet
      description: |
        Create a new GameServerFleet in the specified namespace (admin only).
        The target namespace is auto-created (and hardened) if it does not
        exist. The referenced GameProfile must exist and replicas must be >= 0.
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameServerFleet"
      responses:
        "201":
          description: Fleet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServerFleet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/gameserverfleets/{namespace}/{name}:
    get:
      operationId: getGameServerFleet
      tags: [Fleets]
      summary: Get fleet
      description: Get a specific GameServerFleet
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Fleet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServerFleet"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGameServerFleet
      tags: [Fleets]
      summary: Scale fleet
      description: |
        JSON merge-patch for fleet scaling (operator+ only). Only
        `spec.replicas` may be changed; any other field is rejected with 400.
        replicas must be >= 0.
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameServerFleetPatch"
      responses:
        "200":
          description: Updated fleet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServerFleet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGameServerFleet
      tags: [Fleets]
      summary: Delete fleet
      description: Delete a GameServerFleet and its GameServers (admin only)
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "204":
          description: Fleet deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/apikeys:
    get:
      operationId: listAPIKeys
      tags: [API Keys]
      summary: List API keys
      description: List all API keys without secret values (admin only)
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createAPIKey
      tags: [API Keys]
      summary: Create API key
      description: Generate a new API key (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: API key created (key value shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/apikeys/{keyId}:
    delete:
      operationId: deleteAPIKey
      tags: [API Keys]
      summary: Revoke API key
      description: Delete/revoke an API key (admin only)
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: API key revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    Namespace:
      name: namespace
      in: path
      required: true
      schema:
        type: string
    Name:
      name: name
      in: path
      required: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        OIDC JWT token for humans, or a minato API key (`minato_...`) for
        machine callers. `X-API-Key` is accepted as a legacy alias for API keys.
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication (local development only)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "Legacy alias for API key authentication; prefer `Authorization: Bearer`"
  schemas:
...
    GameServer:
      type: object
      properties:
        apiVersion:
          type: string
          example: operator.minato.io/v1
        kind:
          type: string
          example: GameServer
        metadata:
          $ref: "#/components/schemas/ObjectMeta"
        spec:
          $ref: "#/components/schemas/GameServerSpec"
        status:
          $ref: "#/components/schemas/GameServerStatus"
    ObjectMeta:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        creationTimestamp:
          type: string
          format: date-time
    GameServerSpec:
      type: object
      properties:
        profile:
          type: string
          description: Reference to GameProfile name
        tier:
          type: string
          maxLength: 32
          description: Resource tier from the GameProfile's spec.resources.tiers
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
        storage:
          $ref: "#/components/schemas/StorageSpec"
        priorityClassName:
          type: string
        loadBalancerIP:
          type: string
          description: Pin a specific external IP for the game server's LoadBalancer service
        lifecycle:
          type: object
          properties:
            idleTimeoutSeconds:
              type: integer
            autoStart:
              type: boolean
    GameServerStatus:
      type: object
      properties:
        state:
          type: string
          enum: [Provisioning, Running, Idle, Stopped, Error]
        players:
          type: integer
          description: Current player count
        playerCapacity:
          type: integer
          description: Maximum player capacity
        agentVersion:
          type: string
        endpoints:
          type: array
          items:
            $ref: "#/components/schemas/Endpoint"
        conditions:
          type: array
          items:
            type: object
    Endpoint:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        port:
          type: integer
    StorageSpec:
      type: object
      properties:
        size:
          type: string
          example: 20Gi
        storageClass:
          type: string
        snapshotRef:
          type: object
          properties:
            name:
              type: string
            namespace:
              type: string
...
    ActionExecutionRef:
      type: object
      properties:
        name:
          type: string
          description: Execution ID
      required: [name]
...
    GameSnapshot:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
          example: GameSnapshot
        metadata:
          $ref: "#/components/schemas/ObjectMeta"
        spec:
          type: object
          properties:
            gameServerRef:
              type: string
            schedule:
              type: string
              description: Cron expression for scheduled snapshots
            retention:
              type: object
              properties:
                count:
                  type: integer
                duration:
                  type: string
                  example: 720h
        status:
          type: object
          properties:
            state:
              type: string
              enum: [Pending, Ready, Failed]
            readyAt:
              type: string
              format: date-time
            size:
              type: string
    GameServerFleet:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
          example: GameServerFleet
        metadata:
          $ref: "#/components/schemas/ObjectMeta"
        spec:
          type: object
          properties:
            profile:
              type: string
            replicas:
              type: integer
            template:
              type: object
              properties:
                spec:
                  $ref: "#/components/schemas/GameServerSpec"
            updateStrategy:
              type: object
              properties:
                type:
                  type: string
                  enum: [RollingUpdate, Recreate]
                rollingUpdate:
                  type: object
                  properties:
                    maxUnavailable:
                      type: integer
            priorityClassName:
              type: string
            topologySpreadConstraints:
              type: array
              items:
                type: object
        status:
          type: object
          properties:
            replicas:
              type: integer
            readyReplicas:
              type: integer
            updatedReplicas:
              type: integer
            availableReplicas:
              type: integer
            conditions:
              type: array
              items:
                type: object
...
    APIKey:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [viewer, operator, admin]
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        namespaces:
          type: array
          items:
            type: string
          description: Namespace patterns the key is restricted to (empty = cluster-wide)
        expiresAt:
          type: string
          format: date-time
          description: Optional expiry time; expired keys are rejected with 401
    CreateAPIKeyRequest:
      type: object
      properties:
        name:
          type: string
          description: Key name
        role:
          type: string
          enum: [viewer, operator, admin]
          description: Role for the key (defaults to the caller's role)
        namespaces:
          type: array
          items:
            type: string
          description: >-
            Optional namespace patterns (exact names or trailing-"*" globs such
            as "tenant-*") restricting the key. Omit for a cluster-wide key.
        expiresAt:
          type: string
          format: date-time
          description: Optional expiry time (must be in the future)
      required: [name]
    APIKeyCreated:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        namespaces:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
        key:
          type: string
          description: The API key value (shown once)
        warning:
          type: string
      required: [name, role, key]
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    GameServerLifecycle:
      type: object
      additionalProperties: false
      properties:
        autoStart:
          type: boolean
          description: >-
            Whether the server should be running. false gracefully stops a
            running server (agent PrepareShutdown, scale to 0, state Stopped);
            true starts a stopped server again.
        idleTimeoutSeconds:
          type: integer
          minimum: 0
          description: Seconds of player-idle time before auto-shutdown (0 = never)
    GameServerLifecyclePatch:
      type: object
      additionalProperties: false
      description: >-
        Strict merge-patch for GameServer lifecycle control. Only
        spec.lifecycle.autoStart / spec.lifecycle.idleTimeoutSeconds are
        accepted; unknown fields are rejected with 400.
      properties:
        spec:
          type: object
          additionalProperties: false
          properties:
            lifecycle:
              $ref: "#/components/schemas/GameServerLifecycle"
    GameServerFleetPatch:
      type: object
      additionalProperties: false
      description: >-
        Strict merge-patch for fleet scaling. Only spec.replicas is accepted;
        unknown fields are rejected with 400.
      properties:
        spec:
          type: object
          additionalProperties: false
          properties:
            replicas:
              type: integer
              minimum: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...